Remove success-meta-helper; update delete reducer to be inline
