+++ src/reducers/delete.js
@@ -1,5 +1,4 @@
 import updateMetaHelper from './update-meta-helper';
-import successMetaHelper from './success-meta-helper';
 import requestStatuses from '../utils/request-statuses';
 
 export function del(state, action) {
@@ -35,13 +34,68 @@ export function delNull(state, action) {
 export function delSucceed(state, action) {
   const ids = action.ids;
   const requestLabel = action.requestLabel;
+  let newMeta, newListMeta, newLabels;
 
-  const allMeta = successMetaHelper({
-    crudAction: 'delete',
-    requestLabel,
-    state,
-    ids
-  });
+  const meta = state.meta;
+  const listMeta = state.listMeta;
+  const labels = state.labels;
+
+  const hasIds = ids && ids.length;
+
+  if (requestLabel) {
+    newListMeta = listMeta;
+
+    const existingLabel = state.labels[requestLabel] || {};
+    const existingLabelIds = existingLabel.ids || [];
+
+    let newLabelIds = existingLabelIds.filter(r => !ids.includes(r));
+
+    newLabels = {
+      ...labels,
+      [requestLabel]: {
+        ...existingLabel,
+        ids: newLabelIds,
+        status: requestStatuses.SUCCEEDED
+      }
+    };
+
+    let nullMeta = {};
+    if (ids) {
+      nullMeta = ids.reduce((memo, id) => {
+        memo[id] = null;
+        return memo;
+      }, {});
+    }
+
+    newMeta = {
+      ...meta,
+      ...nullMeta
+    };
+  }
+
+  else if (!hasIds) {
+    newMeta = meta;
+    newListMeta = {
+      ...listMeta,
+      deleteStatus: requestStatuses.SUCCEEDED
+    };
+    newLabels = labels;
+  }
+
+  else {
+    const nullMeta = ids.reduce((memo, id) => {
+      memo[id] = null;
+      return memo;
+    }, {});
+
+    newMeta = {
+      ...meta,
+      ...nullMeta
+    };
+
+    newListMeta = listMeta;
+    newLabels = labels;
+  }
 
   // Shallow clone the existing resource array, removing the deleted resource
   let resources;
@@ -53,9 +107,9 @@ export function delSucceed(state, action) {
 
   return {
     ...state,
-    ...allMeta,
-    // meta: newMeta,
-    // labels,
+    meta: newMeta,
+    labels: newLabels,
+    listMeta: newListMeta,
     resources
   };
 }
