+++ src/reducers/update.js
@@ -1,5 +1,5 @@
 import updateMetaHelper from './update-meta-helper';
-import successMetaHelper from './success-meta-helper';
+// import successMetaHelper from './success-meta-helper';
 import requestStatuses from '../utils/request-statuses';
 import upsertResources from '../utils/upsert-resources';
 
@@ -36,14 +36,14 @@ export function updateNull(state, action) {
 export function updateSucceed(state, action) {
   const resources = action.resources;
   const mergeResources = action.mergeResources;
-  const ids = resources.map(r => r.id);
+  // const ids = resources.map(r => r.id);
 
-  const allMeta = successMetaHelper({
-    requestLabel: action.requestLabel,
-    crudAction: 'update',
-    state,
-    ids
-  });
+  // const allMeta = successMetaHelper({
+  //   requestLabel: action.requestLabel,
+  //   crudAction: 'update',
+  //   state,
+  //   ids
+  // });
 
   const newResources = upsertResources({
     resources: state.resources,
@@ -53,7 +53,7 @@ export function updateSucceed(state, action) {
 
   return {
     ...state,
-    ...allMeta,
+    // ...allMeta,
     resources: newResources
   };
 }
